JCLOUDS-686: Remove changesSince() query from base PaginationOptions
