+++ apis/openstack-keystone/src/main/java/org/jclouds/openstack/v2_0/domain/PaginatedCollection.java
@@ -30,9 +30,6 @@ import static org.jclouds.http.utils.Queries.queryParser;
 /**
  * Base class for a paginated collection in OpenStack.
  *
- * @see <a
- *      href="http://docs.openstack.org/api/openstack-identity-service/2.0/content/Paginated_Collections-d1e325.html">
- *      docs</a>
  */
 public class PaginatedCollection<T> extends IterableWithMarker<T> {
    private final Iterable<T> resources;
